requirements.txt
fastapi
uvicorn
python-multipart
google-cloud-firestore
google-generativeai
PyPDF2
